+++ kutana/backends/__init__.py
@@ -1,6 +1,11 @@
 from .debug import Debug
-from .vkontakte import Vkontakte
+from .vkontakte import Vkontakte, VkontakteLongpoll, VkontakteCallback
 from .telegram import Telegram
 from .terminal import Terminal
 
-__all__ = ["Debug", "Vkontakte", "Telegram", "Terminal"]
+__all__ = [
+    "Debug",
+    "Vkontakte", "VkontakteLongpoll", "VkontakteCallback",
+    "Telegram",
+    "Terminal"
+]