+++ tests/test_backends_api.py
@@ -11,16 +11,15 @@ def test_backend_identities():
     for backend_name in kutana.backends.__all__:
         backend_class = getattr(kutana.backends, backend_name)
         identity = backend_class.get_identity()
-        assert identity not in identities
         identities.add(identity)
+
+    assert identities
 
     for i1 in identities:
         for i2 in identities:
             if i1 == i2:
                 continue
             assert not i1.endswith(i2)
-
-    assert len(identities) == len(kutana.backends.__all__)
 
 
 @patch("kutana.backends.Vkontakte._request")